requirements.txt
django
channels
channels_redis
networkx
websocket_client